+++ src/registrar/assets/sass/_theme/_admin.scss
@@ -271,10 +271,12 @@ h1, h2, h3,
     }
 }
 
+// Hides the "clear" button on autocomplete, as we already have one to use
 .select2-selection__clear {
     display: none;
 }
 
+// Fixes a display issue where the list was entirely white, or had too much whitespace
 .select2-dropdown {
     display: inline-grid !important;
 }
